Updated to Daniel Glassey's new comp drivers
git-svn-id: https://crosswire.org/svn/sword/trunk@84 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ src/modules/common/swcomprs.cpp
@@ -66,7 +66,7 @@ char *SWCompress::Buf(char *ibuf)
 }
 
 
-char *SWCompress::zBuf(unsigned short *len, char *ibuf)
+char *SWCompress::zBuf(unsigned long *len, char *ibuf)
 {
 	if (ibuf) {
 		Init();
